Delete interface folder, keep interface class inside implement class for easy debug
