+++ BusinessLogicLayer/Implements/Services/ProductService.cs
@@ -1,13 +1,17 @@
 ﻿using AutoMapper;
 using BusinessLogicLayer.DTO.ProductDTO;
 using BusinessLogicLayer.Implements.Base;
-using BusinessLogicLayer.Interfaces.Services;
 using DataAccessLayer.Models.ProductEntities;
 using DataAccessLayer.Repository.IRepository.Base;
 using Microsoft.AspNetCore.Http;
 
 namespace BusinessLogicLayer.Implements.Services
 {
+    public interface IProductService : ICrudService<ProductCreateDTO, ProductGetDTO, ProductUpdateDTO, Product>
+    {
+        public Task<Product> CreateProductWithImageAsync(Product product, IFormFile imageFile);
+    }
+
     public class ProductService(IUnitOfWork _unitOfWork, IImageService _imageService, IMapper _mapper) : CrudService<ProductCreateDTO, ProductGetDTO, ProductUpdateDTO, Product>(_unitOfWork, _mapper, _imageService), IProductService
     {
