+++ BusinessLogicLayer/Implements/Services/UserService.cs
@@ -6,6 +6,11 @@ namespace BusinessLogicLayer.Implements.Services
 {
     public class UserService(IUnitOfWork _unitOfWork) : IUserService
     {
+        public interface IUserService
+        {
+            public Task<User> GetUserById(int id);
+        }
+
         public async Task<User> GetUserById(int id)
         {
             return await _unitOfWork.Repository<User>().GetByIdAsync(id);
