--- BusinessLogicLayer/Interfaces/Services/IAuthService.cs
@@ -1,11 +1,0 @@
-﻿using BusinessLogicLayer.DTO.UserDTO;
-using DataAccessLayer.Models.UserEntities;
-
-namespace BusinessLogicLayer.Interfaces.Services
-{
-    public interface IAuthService
-    {
-        public Task<User> RegisterAsync(string email, string password);
-        public Task<AuthUserRespondDTO> AuthenticateAsync(string email, string password, string ipAddress);
-    }
-}
