--- BusinessLogicLayer/Interfaces/Services/IImageService.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-
-namespace BusinessLogicLayer.Interfaces.Services
-{
-    public interface IImageService
-    {
-        public Task<string> UploadImageAsync(IFormFile file, string bucketName);
-    }
-}
