--- BusinessLogicLayer/Interfaces/Services/IProductCategoryService.cs
@@ -1,6 +1,0 @@
-﻿namespace BusinessLogicLayer.Interfaces.Services
-{
-    public interface IProductCategoryService
-    {
-    }
-}
