--- BusinessLogicLayer/Interfaces/Services/IUserService.cs
@@ -1,9 +1,0 @@
-﻿using DataAccessLayer.Models.UserEntities;
-
-namespace BusinessLogicLayer.Interfaces.Services
-{
-    public interface IUserService
-    {
-        public Task<User> GetUserById(int id);
-    }
-}
