--- BusinessLogicLayer/Interfaces/Services/IVoucherService.cs
@@ -1,10 +1,0 @@
-﻿using BusinessLogicLayer.DTO.VoucherDTO;
-using BusinessLogicLayer.Interfaces.Base;
-using DataAccessLayer.Models.UserEntities;
-
-namespace BusinessLogicLayer.Interfaces.Services
-{
-    public interface IVoucherService : ICrudService<VoucherCreateDTO, VoucherGetDTO, VoucherUpdateDTO, Voucher>
-    {
-    }
-}
