--- DataAccessLayer/Repository/IRepository/Base/IUnitOfWork.cs
@@ -1,10 +1,0 @@
-﻿using DataAccessLayer.Models.AbstractEntities;
-
-namespace DataAccessLayer.Repository.IRepository.Base
-{
-    public interface IUnitOfWork
-    {
-        public Task<int> SaveChangesAsync();
-        IGenericRepository<T> Repository<T>() where T : BaseEntity;
-    }
-}
